Create visual generalization
